* Source/art/blit.m(MPRE): Remove compiler warning.         * Source/art/image.m(-DPSimage:::::::::::): Correct parameter         types. Fixes bug #41274.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@37615 72102866-910b-0410-8b05-ffd578937521

+++ Source/art/blit.m
@@ -318,7 +318,7 @@ static void MPRE(run_opaque) (render_run_t *ri, int num)
 
   COPY_ASSEMBLE_PIXEL(v, ri->r, ri->g, ri->b)
   v = v + (v << 16);
-  if (((int)dst&2) && num)
+  if (((NSUInteger)dst&2) && num)
     {
       *dst++ = v;
       num--;
